Fix warnings on gcc 11.5

+++ src/programs/unit_tests/Components/BarrierEventTests.cc
@@ -8,7 +8,7 @@
 #include "JANA/Utils/JBenchUtils.h"
 #include "catch.hpp"
 
-int global_resource = 0;
+size_t global_resource = 0;
 
 
 struct BarrierSource : public JEventSource {
